Refs #124. Moved direct formatting bold testing to its own module
