+++ tests/export/test_html.py
@@ -770,121 +770,6 @@ class StyleBasedOnTestCase(DocumentGeneratorTestCase):
         self.assert_document_generates_html(document, expected_html)
 
 
-class DirectFormattingBoldPropertyTestCase(DocumentGeneratorTestCase):
-    def test_default_no_val_set(self):
-        document_xml = '''
-            <p>
-              <r>
-                <rPr>
-                  <b />
-                </rPr>
-                <t>foo</t>
-              </r>
-            </p>
-        '''
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '<p><strong>foo</strong></p>'
-        self.assert_document_generates_html(document, expected_html)
-
-    def test_valid_enable_vals_create_strong(self):
-        vals = [
-            'true',
-            'on',
-            '1',
-            '',
-        ]
-        paragraph_template = '''
-            <p>
-              <r>
-                <rPr>
-                  <b val="%s" />
-                </rPr>
-                <t>foo</t>
-              </r>
-            </p>
-        '''
-        document_xml = ''.join(
-            paragraph_template % val
-            for val in vals
-        )
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '''
-            <p><strong>foo</strong></p>
-            <p><strong>foo</strong></p>
-            <p><strong>foo</strong></p>
-            <p><strong>foo</strong></p>
-        '''
-        self.assert_document_generates_html(document, expected_html)
-
-    def test_valid_disabled_vals_do_not_create_strong(self):
-        vals = [
-            'off',
-            'false',
-            'none',
-            '0',
-        ]
-        paragraph_template = '''
-            <p>
-              <r>
-                <rPr>
-                  <b val="%s" />
-                </rPr>
-                <t>foo</t>
-              </r>
-            </p>
-        '''
-        document_xml = ''.join(
-            paragraph_template % val
-            for val in vals
-        )
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '''
-            <p>foo</p>
-            <p>foo</p>
-            <p>foo</p>
-            <p>foo</p>
-        '''
-        self.assert_document_generates_html(document, expected_html)
-
-    def test_invalid_vals_do_not_create_strong(self):
-        vals = [
-            'foo',
-            'bar',
-        ]
-        paragraph_template = '''
-            <p>
-              <r>
-                <rPr>
-                  <b val="%s" />
-                </rPr>
-                <t>foo</t>
-              </r>
-            </p>
-        '''
-        document_xml = ''.join(
-            paragraph_template % val
-            for val in vals
-        )
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '''
-            <p>foo</p>
-            <p>foo</p>
-        '''
-        self.assert_document_generates_html(document, expected_html)
-
-
 class DrawingGraphicBlipTestCase(DocumentGeneratorTestCase):
     def test_inline_image_with_multiple_ext_definitions(self):
         # Ensure that the image size can be calculated correctly even if the
